backend/requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.75.2
fastapi-admin==1.0.3
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
orjson==3.6.8
packaging==21.3
pendulum==2.1.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyparsing==3.0.8
pypika-tortoise==0.1.4
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.0
zipp==3.8.0

pytest
requests


#for  cleaning up codebase on each commit
pre-commit

#seedir to print tree structure of folders
seedir
seedir

